requirements.txt
# api deps
sqlalchemy[asyncio]
alembic
asyncpg
fastapi
fastapi-cli
pydantic<2.9.0
python-dotenv
uvicorn
pydantic_settings
python-jose[cryptography]
passlib[bcrypt]
pyjwt
python-multipart
alembic-postgresql-enum
# bot deps
aiogram
python-decouple
prettytable
redis
openpyxl